Add deleteFileWithRetries
For some reason, after all the image processing is completed and the worker pool has been joined, the application can still fail the delete the source images because another process is still using the file. I suspect that Python is not waiting for the OS to fully terminate the subprocesses it launched, or the OS is slow in cleaning up process resources like file handles. Since I can't get into the guts of what Python's multiprocess library is doing, I'm just throwing a 5 count retry loop with a 0.5 second delay on each failed attempt on the problematic delete in the hopes that this solves most cases.