Ignoring missing image metadata (#850)
During conversion, the code expected that every processed file has a
metadata object attached in imgMetadata dictionary. For some reason this
is not true for every file.
I've added code that adds empty metadata to every file where the
metadata is missing.
This is more of a workaround, so the conversion can finish, it would be
better to find out why the metadata is missing in the first place, but I
don't see that.
Since most images have empty metadata, I don't think we are missing any
information.

This fixes #850